requirements.txt
pytest==6.2.5
requests==2.26.0
python-dotenv==0.19.2